chore: expose articles request meta data
